Refactor economy data handling and remove unused exception
- Replaced `double` with `long` for financial fields across economy-related classes to improve precision and consistency.
- Removed unused `TooRichException` from project.
- Enhanced `DataManager`, `BankManager`, and `CharacterManager` for better caching and data retrieval support.
- Updated `Minty` to periodically save banks asynchronously and added improvements in configuration management.
- Added support for loan management within `BankManager`.

+++ src/main/java/dev/vayen/mc/economy/bank/Bank.java
@@ -34,7 +34,7 @@ public class Bank {
     List<BankCustomer> customers;
     @BsonIgnore
     List<BankLoan> loans;
-    double maxDebt;
+    long maxDebt;
 
     public void addCustomer(BankCustomer customer) {
         customers.add(customer);
